requirements.txt
airtable==0.3.1
airtable-python-wrapper==0.8.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.7.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2017.11.5
cffi==1.14.0
channels==2.3.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cssselect2==0.3.0
cupshelpers==1.0
daphne==2.5.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.0rc1
diff-match-patch==20181111
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.0.3
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-import-export==2.2.0
django-paypal==1.0.0
django-sass==1.0.0
django-sass-processor==0.8
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.6
importlib-metadata==1.5.0
incremental==17.5.0
jdcal==1.4.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libsass==0.20.0
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupPy==1.14
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
mysqlclient==1.4.4
netifaces==0.10.4
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.4
paramiko==2.6.0
pbr==5.4.5
pep517==0.8.2
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
Pyphen==0.9.5
pyRFC3339==1.1
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2017.3
pyxdg==0.26
PyYAML==5.3.1
rcssmin==1.0.6
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rjsmin==1.1.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.48.0
system-service==0.3
systemd-python==234
tablib==2.0.0
testresources==2.0.1
tinycss2==1.0.2
tmdbsimple==2.1.0
toml==0.10.1
traceback2==1.4.0
Twisted==20.3.0
txaio==20.4.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.22
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.1.0
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==1.0.0
zope.interface==5.1.0
